+++ src/main/java/net/danielthompson/danray/cameras/CameraSettings.java
@@ -21,17 +21,17 @@ public class CameraSettings {
    /**
     * Focal length of the simulated lens. Must be >= 0. 0 is isometric; > 0 is perspective. Required.
     */
-   public double FocalLength;
+   public float FocalLength;
 
    /**
     * Rotation about the view vector in degrees (positive = clockwise). Optional; defaults to 0.
     */
-   public double Rotation;
+   public float Rotation;
 
    /**
     * How much to shrink or enlarge the viewport. Optional; defaults to 1; less than 1 = zoomed in; greater than 1 = zoomed out.
     */
-   public double ZoomFactor;
+   public float ZoomFactor;
 
    /**
     * Camera's origin and direction in world space. Required.
@@ -51,6 +51,6 @@ public class CameraSettings {
    /**
     * Distance in world coordinates to the focus point. >0 required for DOF, otherwise optional.
     */
-   public double FocusDistance;
+   public float FocusDistance;
 
 }
