+++ src/main/java/net/danielthompson/danray/cameras/apertures/Aperture.java
@@ -6,6 +6,6 @@ import net.danielthompson.danray.structures.Point;
  * Created by daniel on 3/2/14.
  */
 public abstract class Aperture {
-   public double Size;
+   public float Size;
    public abstract Point GetOriginPoint();
 }
